Add files via upload
This repository provides an overview of some Machine Learning and Deep Learning methods as fundamental tools in detecting potential failures of physical assets using condition monitoring techniques, for this, in the first part supervised learning algorithms are applied for classification and regression in different case studies; comparing the performance of models demonstrates the effectiveness of deep neuronal networks LSTM, whose properties are of great value in sequential data processing and promise more powerful applications in maintenance engineering. In the second part effectiveness is argued by optimally adjusting the neural network architecture and implementing hybrid models that maximize model performance. In the third part describes and implements a Web application to put in production a model of classification of failures in bearings, the algorithm selected for the Web solution is Gradient Boosting due to the good performance with the data set and efficiency in the use of computational resources, with this development the end user access to the classification model is improved. Finally, a survival analysis method is applied with a statistical estimator, the purpose of which is to calculate the average life of the machine and the survival curves to compare the probability of failure during the time of operation of the physical asset.